Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,24 +1,24 @@
 #include <stdio.h>
 
 int main() {
-    int a,b,z,y,x,w;
+    double a,b,z,y,x,w;
     char c;
-    scanf("%d %d %c",&a,&b,&c);
+    scanf("%lf %lf %c",&a,&b,&c);
     if(c=='+'){
         z = a+b;
-        printf("%d",z);
+        printf("%lf",z);
     }
     else if(c=='-'){
         y = a-b;
-        printf("%d",y);
+        printf("%lf",y);
     }
     else if(c=='*'){
         x = a*b;
-        printf("%d",x);
+        printf("%lf",x);
     }
     else if(c=='/'){
         w = a/b;
-        printf("%d",w);
+        printf("%lf",w);
     }
     return 0;
 }
